+++ clickhouse-data/src/main/java/com/clickhouse/data/stream/NonBlockingPipedOutputStream.java
@@ -15,6 +15,8 @@ import com.clickhouse.data.ClickHouseOutputStream;
 import com.clickhouse.data.ClickHousePipedOutputStream;
 import com.clickhouse.data.ClickHouseUtils;
 import com.clickhouse.data.ClickHouseWriter;
+import com.clickhouse.logging.Logger;
+import com.clickhouse.logging.LoggerFactory;
 
 /**
  * A combination of {@link java.io.PipedOutputStream} and
@@ -23,6 +25,9 @@ import com.clickhouse.data.ClickHouseWriter;
  * reader are on two separate threads.
  */
 public class NonBlockingPipedOutputStream extends ClickHousePipedOutputStream {
+
+    private static final Logger log = LoggerFactory.getLogger(NonBlockingPipedOutputStream.class);
+
     protected final AdaptiveQueue<ByteBuffer> queue;
 
     protected final int bufferSize;
@@ -176,6 +181,7 @@ public class NonBlockingPipedOutputStream extends ClickHousePipedOutputStream {
         ByteBuffer b = buffer;
         while (length > 0) {
             int remain = b.remaining();
+            log.debug("writeBytes length:[%d] remain:[%d] offset: [%d]", length, remain, offset);
             if (length < remain) {
                 b.put(bytes, offset, length);
                 length = 0;